Switch to argument-based accessor for `last_match`
When calling `last_match` and not passing an index to the method, it
returns all of the matchers. This was being used to access the matcher
based on the index array. However, if there is a problem and last_match
returns `nil`, then the array accessor raises an error. Passing an
argument to `last_match` returns nil if no result is present at that
position.