Allow loopback client to modify ntp configuration
Consider a situation where the host runs a local nameserver like BIND or
a proxy like dnsmasq, and ntpd starts before the namserver or proxy
does.  This causes a race condition where ntpd may not be able to
resolve its configured nameserver addresses immediately.

The `nomodify` option added by the cookbook made it so that ntpd is
unable to self-heal - it will not attempt to re-resolve the hosts after
startup.  Removing this option is safe for the loopback interfaces as it
will not introduce a security risk -- and `nomodify` is explicitly
excluded from the loopback interface restriction set on both Ubuntu and
EL (Red Hat/CentOS) distributions.

See also https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=571469 for
additional background.