Add support for ntp `pool` configuration option
This adds support for the ntp `pool` configuration option. It's a simpler way to pull in pools of servers, especially ones that don't provide individual dns records.

I've tried to follow local style. I did not add anything to test kitchen, as I'm unsure of whether it will work on AIX. But adding to the ubuntu ones, and testing locally does work.

Signed-off-by: seph <seph@directionless.org>